Add AJAX files uploads for newslike content
Added drag and drop AJAX uploads for preview images for Posts and
Events, similar to Pages.

Added preview images drop zones in the admin_edit views. In the
controller we added the admin_ajaxUploadFiles method. Since we can
drop arbitrary images for the images, we just upload them
and not do any conversions or resize them on upload. The drop zones
need to have the content year specified in the data-options
attribute. We also setup the beforeFilter() callbacks.