Add functionality to upload several Album images
Added the functionality to upload several Album images and also create
their thumbnails. In the admin_add and admin_edit views we add a new
file upload field for multiple files.

In the controller first we unlock this field in the beforeFilter()
callback method for these actions, since the Security Component
doesn't seem to like them. Then in the actions we add the
uploadFiles() call from the Upload component and we set the
create_thumb option to true to create the thumbnails. The file name
parameter is false so the image names are preserved.