Add drop zones for Albums images
We created three divs in the admin_edit view, below the form so it's
clear it's separate from the form submission. We add a class of
dropzone and an id of dropzone-preview-images, dropzone-album and
dropzone-album-thumbnails we can target. Wel also add the album year
as an option for the drop zones. We also add the the album ids to
the event handleres in the admin Javascript file.