Add AJAX upload functionality for Albums images
Added the functionality for AJAX uploading for Albums images via a
drag and drop area. We just duplicated the functionality we did
for the files uploads and adjusted it.

In the controller we added the admin_ajaxUploadFiles method to handle
all upload cases: preview images, album images and thumbnail images.
Since we can drop arbitrary images, we just upload them and not do
any conversions or resize them on upload. The drop zones need to have
the Album year specified in the data-options attribute.