Delete Albums images and preview more generically
Modified the Albums system so we can delete both the images and its
thumbnails, and the preview images, without having to create
different controller methods or add too many parameters.

First in the model we simplified the deleteFile method so that it
only takes a path parameter. This path will be built in the
controller. We don't use the id or other parameters to build the path
since these are built in the controller. We just delete the file and
return the result. Since this methos is very simply and could be
applied to other models, we could probably turn it into a trait when
we upgrade the site to CakePHP v3.0.

In the controller we renamed the admin_deleteAlbumImage method to
admin_deleteFile. The image parameter is renamed filename and we add
a parameter fileType (with default value preview-image) to specify if
we're deleteing preview images or Album images.

After checking the album exists, we retrieve it. We then build the
path of the file to be deleted based on the file type to delete:
preview image or Album image with its thumbnail. We then just call
the deleteFile method of the Album model with the path(s). Based on
the result we set the Flash messages. Finally dependeing if we have
a redirection named parameter we redirect to the appropiate action.
We also fixed a few lines, like using empty() to check parameters and
change the or operator to ||.

In  the views we slightly modified the postLink call. Mainly we pass
the file name as a parameter and renamed the named parameter
redirect_action to redirection. In the admin_view view we also
renamed the action to use from admin_deleteAlbumImage to
admin_deleteFile.