Fix Albums to use query strings to delete images
Fixed the Albums controller and admin_view and admin_edit views to
delete album and preview images. The system was using named
parameters, which are very CakePHP specific, non standard, and marked
for deprecation in CakePHP v3.0.

With this we changed the admin_deleteFile method, now it only takes
the id as parameter, and the filename, fileType and redirection
parameters are passed via query strings. We use the query() call
instead of the property, which is safer because it returns a null
value on non existent parameters.

In the view we change the delete postLink call to use the query
strings, which are specified by using a URL route with key '?' that
specifies an array with key value pairs specifying the query string
parameters.