Fix checking empty values for return function values
Fixed the admin_deleteFile methods for the Albums, Posts and Events,
which were checking the return value of method calls with empty()
calls. This worked in the development environment, which is PHP 5.5.
But this fails in prior versions, which is what we have in the
production site.

In the method we create a filename variable near the top of the method,
since we use it in several places anyway, and assign it the query
value. We delete the same line that was present lower in the method.
We do the same thing with the redirection query string variable.