Delete files of Pages using query strings
Fixed the way the images are deleted in Pages, so they are done in a
more generic manner and using query string parameters, similar to
what we did in the Albums and newslike content systems. We modified
the models, controller and admin_view and admin_edit views of each
system. The system was using named parameters, which are very CakePHP specific, non standard, and marked for deprecation in CakePHP v3.0.

First in the model we simplified the deleteFile method so that it
only takes a path parameter. This path will be built in the
controller. We just delete the file and return the result. Since this
method is very simply and can be applied to other models, we could
probably turn it into a trait when we upgrade the site to CakePHP
v3.0.

We changed the admin_deleteFile method, now it only takes the id as
parameter, and the filename and redirection parameters are passed via
query strings. For the Pages and Projects system we also have a
fileType parameter to select which specific type of file, for example
images or documents in Pages we want to delete. We use the query()
call instead of the property, which is safer because it returns a null
value on non existent parameters.

After checking the record exists, we retrieve it. We then just call
the deleteFile method of the model with the path. Based on
the result we set the Flash messages. Finally depending if we have
a redirection parameter we redirect to the appropiate action.

In the views we change the delete postLink call to use the query
strings, which are specified by using a URL route with key '?' that
specifies an array with key value pairs specifying the query string
parameters.