Create paginateCount method for Album Image model
Created an overridden paginateCount method in the AlbumImage model
since pagination with the Files List Datasource was not working. It
would just display the first page of images, based on the limit.
The paginateCount method just gets the list of files, based on the
folder condition, and counts it.

TODO: figure out why the datasource is not working with pagination
like before.