Fix Files List Datasource to return formatted results
Fixed the FilesListSource so that it returns results in a format
similar to regular find operations. This also fixes the pagination
problem in the previous commit.

Previously we return the results with top level key, the model alias,
which then contained an array of all the files. Now we return an array
with numeric keys, and each key contains an array with the model alias,
that is a key to an array containing a single file name. This is
similar to using a find('all') operation. We also modified the view
view accordingly.

Since the pagination issue is fixed we can change the Album Image
model to the way it was before. We don't need to include the
overridden paginateCount method.