Add Albums images to admin_edit view
Album images are now displayed in the admin_edit view so we can see
the thumbnails and delete images if necessary, similar to the
admin_view view. In the controller we just call the getAlbumThumbnails
method, using the request data to get the needed parameters.

In the view we display all thumbnails above the preview images.