Update to CakePHP v2.6.x
Updated base files changed between CakePHP v2.5.8 and v2.6.1. Some
changes were minor, like comments in the database.php and acl.php
files. In bootstrap.php we have a commented out line about the way
to do translations (i18).

The important changes are in the index.php and test.php files in the
webroot, which includes a couple of lines to detect if CakePHP is
installed via Composer, which is the preferred way in CakePHP v3.0.