Add type parameter to parentNode methods
Added the type parameter to the parentNode methods in the User and
Group models- This is a new feature of CakePHP v2.6.x: when using
AclBehavior, The model's parentNode() method now gets the type
(Aro, Aco) passed as first argument.

Since our ACL system is simple, we don't use this parameter in the
methods, it doesn't affect hte behavior.