Update of package.json file
Updated the package.json file used in the frontend development. The
new Bootstrap v3.3.2 introduced a few changes, mainly the use of the
CSS autoprefixer system.

Revieweing the changes revealed that the official Bootstrap upgraded
the Grunt plugins we used, so we also did the same for the frontend
development. We also added the autoprefixer Grunt plugin to the list
of packages.

We glogablly installed the npm-check-updates Node tool which makes
upgrading the packages more easy. npm-check-updates is a tool that
allows you to find the latest versions of dependencies, regardless
of any version constraints in your package.json file (unlike npm
itself). More information in the following links:

https://www.npmjs.com/package/npm-check-updates
http://stackoverflow.com/questions/16525430/npm-check-and-update-package-if-needed