Upgrade to Bootstrap v3.3.2
Added the customized Bootstrap configuration and CSS files
corresponding to v3.3.2. This customized Bootstrap files are used
to generate the compiled project files.

Custom variables remain the same from the v3.3.x versions:

@heading-font-family: 'Open Sans', Helvetica, Arial, sans-serif;

The configuration has been saved to:

https://gist.github.com/2d42736f0c3e6d353b05

and can be revisited at:

http://getbootstrap.com/customize/?id=2d42736f0c3e6d353b05

for further customization.

There was a problem with the previous two Bootstrap versions commit
messages, where we incorrectly copied the message from the ILMfan
project.