Fix delete files in Pages Sections
Fixed the buttons to delete files within the admin_view view of the
Pages Sections. It used the previous delete files methods from the
Pages controller, which now use query strings.

In the views we only pass the id field of the parent model, Page
(the one in the belongsTo relationship), as a parameter and
specify the file name, file type and redirection, for the file in
question. This are turned into query strings using the '?' key which
has an array of key/value pairs specifying the query string
parameters.