Fix paths for Grunt copy task
Fixed the path of the Grunt copy task, since it was still using the
older cms (VFXfan) path to copy the CSS and Javascript files to the
webroot directory.

The paths to generate the API docs, via apigen, have also been fixed,
but the task needs to be fixed to work with a recent install of
apigen.