Upgrade to Bootstrap v3.3.4
Added the customized Bootstrap configuration and CSS files
corresponding to v3.3.4. This customized Bootstrap files are used
to generate the compiled project files.

Custom variables remain the same from the v3.3.x versions:

@heading-font-family: 'Open Sans', Helvetica, Arial, sans-serif;

The configuration has been saved to:

https://gist.github.com/e7331d53d88020f3c9dd

and can be revisited at:

http://getbootstrap.com/customize/?id=e7331d53d88020f3c9dd

for further customization.

Also fixed a few of the  docblock comments, mainly the copyright
info on the Javascript files.